+++ lbh15/properties/bismuth_thermochemical_properties/bismuth_thermochemical.py
@@ -8,6 +8,7 @@ from scipy.constants import R
 from ..interface import PropertyInterface
 from ..common_interface import OxygenPartialPressureInterface
 from ..common_interface import MolarEnthalpyInterface
+from ..common_interface import MolarEntropyInterface
 from ..bismuth_properties import h
 from ..._commons import BISMUTH_BOILING_TEMPERATURE as T_b0
 from ..._commons import BISMUTH_MELTING_TEMPERATURE as T_m0
@@ -303,7 +304,7 @@ class MolarEnthalpy(MolarEnthalpyInterface):
         return f"{self.long_name} in liquid bismuth"
 
 
-class MolarEntropy(PropertyInterface):
+class MolarEntropy(MolarEntropyInterface):
     """
     Liquid bismuth *molar entropy variation* property class.
     """
@@ -335,33 +336,12 @@ class MolarEntropy(PropertyInterface):
                            - 7.183e6 / 2 * (1 / T / T - 1 / T_m0 / T_m0))
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "S"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Molar entropy variation unit
-        """
-        return "[J/(mol.K)]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the molar entropy
         variation correlation function
         """
         return [T_m0, T_b0]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Molar entropy variation long name
-        """
-        return "molar entropy variation"
 
     @property
     def description(self) -> str:
